refactor(acl): split ControlGroupsController into 5 single-action controllers
Each controller now does exactly one thing (SOLID / SRP):

- ShowControlGroupsController  → GET /          (acl.groups)
- CreateControlGroupController → POST /create   (acl.create)
- EditControlGroupController   → GET /acl/{id}  (acl.edit)
- UpdateControlGroupController → POST /acl/{id} (acl.update)
- SearchAffiliatableController → GET /search    (acl.search.affiliatable)

Redirects in create/update now use route() instead of action() so they
no longer depend on ControlGroupsController.

Delete ControlGroupsController.php.

Update routes/Routes/AccessControl/View.php to use new controllers.

Split ControlGroupsControllerTest into 5 matching test files:
- ShowControlGroupsControllerTest    (3 tests)
- CreateControlGroupControllerTest   (3 tests)
- EditControlGroupControllerTest     (3 tests)
- UpdateControlGroupControllerTest   (4 tests)
- SearchAffiliatableControllerTest   (4 tests)

Fix RouteTest: 'does not protect acl routes' was incorrect after the
view access control middleware was added to acl.groups.

253 tests pass.
