+++ tests/Feature/RouteTest.php
@@ -6,8 +6,8 @@ it('protects configurations routes', function () {
         ->assertForbidden();
 });
 
-it('does not protect access control routes', function () {
-    $response = test()->actingAs(test()->test_user)
+it('access control routes require view access control permission', function () {
+    test()->actingAs(test()->test_user)
         ->get(route('acl.groups'))
-        ->assertOk();
+        ->assertForbidden();
 });
